+++ src/__tests__/BigPlayButton.spec.js
@@ -8,6 +8,7 @@ describe('BigPlayButton', () => {
       <BigPlayButton
         player={{
           hasStarted: false,
+          currentSrc: 'https://media.w3.org/2010/05/sintel/trailer_hd.mp4'
         }}
       />
     );
@@ -20,6 +21,7 @@ describe('BigPlayButton', () => {
       <BigPlayButton
         player={{
           hasStarted: true,
+          currentSrc: 'https://media.w3.org/2010/05/sintel/trailer_hd.mp4'
         }}
       />
     );
@@ -32,9 +34,23 @@ describe('BigPlayButton', () => {
       <BigPlayButton
         player={{
           hasStarted: false,
+          currentSrc: 'https://media.w3.org/2010/05/sintel/trailer_hd.mp4'
         }}
       />
     );
     expect(wrapper.hasClass('video-react-big-play-button')).toBe(true);
+  });
+
+  it('should render with "video-react-big-play-button-center" class with center position', () => {
+    const wrapper = shallow(
+      <BigPlayButton
+        position="center"
+        player={{
+          hasStarted: false,
+          currentSrc: 'https://media.w3.org/2010/05/sintel/trailer_hd.mp4'
+        }}
+      />
+    );
+    expect(wrapper.hasClass('video-react-big-play-button-center')).toBe(true);
   });
 });